Work
Removed the toggle button that was at the top of the play screen. It was just for quick debug anyways.
Added colorpicker dialog
Added starting player dialog
Added mask when starting player dialog and colorpicker dialog are opened
Instead of automatically opening starting player dialog when newGame is called I moved it to a button in the menu.
Prevent text selection.
Refactored show / hide code