Fixes
Fix Firefox and iOS issue. They were choking on the member variable initialization of the Player class internals. Weird. Just removed the initializers.
Changed Hammer.js initializer to handle only the pressup and not the tap to prevent the player life total from changing when trying to open the colorpicker.
Clear focus from colorpicker else there are two swatches that display focus: the default color and a changed color. if one exists. Just blur the swatches when colorpicker is closed.